Fixed:Refactor linear regression implementation from scikit-learn to scipy
Due to the unavailability of the scikit-learn module on Cirrus (jobID=5524709) and subsequent issues with numpy module dependencies when using Anaconda, this update refactors the linear regression implementation to utilize scipy, which is supported natively without requiring additional complex modules. This change aligns with Cirrus's Python documentation recommendations and ensures compatibility and stability of the code on the platform.